+++ backend/beverage-service/src/main/java/com/caffeinedoctor/beverageservice/contoller/BeverageController.java
@@ -1,0 +1,14 @@
+package com.caffeinedoctor.beverageservice.contoller;
+
+import org.springframework.cloud.client.discovery.EnableDiscoveryClient;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.RestController;
+
+@RestController
+@EnableDiscoveryClient
+public class BeverageController {
+    @GetMapping("/")
+    public String getServiceName(){
+        return  "beverage-service";
+    }
+}
