--- backend/beverage-service/src/main/java/com/mst/beverageservice/BeverageServiceApplication.java
@@ -1,15 +1,0 @@
-package com.mst.beverageservice;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.cloud.client.discovery.EnableDiscoveryClient;
-
-@SpringBootApplication
-@EnableDiscoveryClient
-public class BeverageServiceApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(BeverageServiceApplication.class, args);
-	}
-
-}
