--- backend/beverage-service/src/main/java/com/mst/beverageservice/contoller/BeverageController.java
@@ -1,12 +1,0 @@
-package com.mst.beverageservice.contoller;
-
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.RestController;
-
-@RestController
-public class BeverageController {
-    @GetMapping("/")
-    public String getServiceName(){
-        return  "beverage-service";
-    }
-}
