+++ backend/beverage-service/src/main/resources/application.yml
@@ -5,7 +5,7 @@ spring:
   application:
     name: beverage-service
 
-# ??? ??? ??
+# 유레카 서버 설정
 eureka:
   client:
     register-with-eureka: true
