+++ backend/beverage-service/src/test/java/com/caffeinedoctor/beverageservice/BeverageServiceApplicationTests.java
@@ -1,0 +1,13 @@
+package com.caffeinedoctor.beverageservice;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class BeverageServiceApplicationTests {
+
+	@Test
+	void contextLoads() {
+	}
+
+}
