--- backend/beverage-service/src/test/java/com/mst/beverageservice/BeverageServiceApplicationTests.java
@@ -1,13 +1,0 @@
-package com.mst.beverageservice;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class BeverageServiceApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
