+++ backend/community-service/src/main/java/com/caffeinedoctor/communityservice/CommunityServiceApplication.java
@@ -1,0 +1,15 @@
+package com.caffeinedoctor.communityservice;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.cloud.client.discovery.EnableDiscoveryClient;
+
+@SpringBootApplication
+@EnableDiscoveryClient
+public class CommunityServiceApplication {
+
+	public static void main(String[] args) {
+		SpringApplication.run(CommunityServiceApplication.class, args);
+	}
+
+}
