+++ backend/community-service/src/main/java/com/caffeinedoctor/communityservice/controller/CommunityController.java
@@ -1,0 +1,13 @@
+package com.caffeinedoctor.communityservice.controller;
+
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.RestController;
+
+@RestController
+public class CommunityController {
+
+    @GetMapping("/")
+    public String getServiceName(){
+        return "community-service";
+    }
+}
