--- backend/community-service/src/main/java/com/mst/communityservice/CommunityServiceApplication.java
@@ -1,15 +1,0 @@
-package com.mst.communityservice;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.cloud.client.discovery.EnableDiscoveryClient;
-
-@SpringBootApplication
-@EnableDiscoveryClient
-public class CommunityServiceApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(CommunityServiceApplication.class, args);
-	}
-
-}
