--- backend/community-service/src/main/java/com/mst/communityservice/controller/CommunityController.java
@@ -1,13 +1,0 @@
-package com.mst.communityservice.controller;
-
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.RestController;
-
-@RestController
-public class CommunityController {
-
-    @GetMapping("/")
-    public String getServiceName(){
-        return "community-service";
-    }
-}
