+++ backend/community-service/src/main/resources/application.yml
@@ -5,7 +5,7 @@ spring:
   application:
     name: community-service
 
-# ??? ??? ??
+# 유레카 설정
 eureka:
   client:
     register-with-eureka: true
