+++ backend/community-service/src/test/java/com/caffeinedoctor/communityservice/CommunityServiceApplicationTests.java
@@ -1,0 +1,13 @@
+package com.caffeinedoctor.communityservice;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class CommunityServiceApplicationTests {
+
+	@Test
+	void contextLoads() {
+	}
+
+}
