--- backend/community-service/src/test/java/com/mst/communityservice/CommunityServiceApplicationTests.java
@@ -1,13 +1,0 @@
-package com.mst.communityservice;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class CommunityServiceApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
