+++ backend/eureka-server/build.gradle
@@ -4,7 +4,7 @@ plugins {
 	id 'io.spring.dependency-management' version '1.1.4'
 }
 
-group = 'com.mst'
+group = 'com.caffeine-doctor'
 version = '0.0.1-SNAPSHOT'
 
 java {
