--- backend/eureka-server/src/main/java/com/mst/eurekaserver/EurekaServerApplication.java
@@ -1,15 +1,0 @@
-package com.mst.eurekaserver;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.cloud.netflix.eureka.server.EnableEurekaServer;
-
-@SpringBootApplication
-@EnableEurekaServer // // EurekaServer로 등록
-public class EurekaServerApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(EurekaServerApplication.class, args);
-	}
-
-}
