+++ backend/eureka-server/src/main/resources/application.yml
@@ -7,7 +7,7 @@ spring:
 
 eureka:
   client:
-    # ??? ??? ????? ???? ?? ??
+    # 유레카 서버가 자기자신을 등룍하는 것을 막음
     register-with-eureka: false
-    # ??? ??? ??? ???? ??
+    # 유레카 서버의 정보를 가져올지 여부
     fetch-registry: false
