+++ backend/eureka-server/src/test/java/com/caffeinedoctor/eurekaserver/EurekaServerApplicationTests.java
@@ -1,0 +1,13 @@
+package com.caffeinedoctor.eurekaserver;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class EurekaServerApplicationTests {
+
+	@Test
+	void contextLoads() {
+	}
+
+}
