--- backend/eureka-server/src/test/java/com/mst/eurekaserver/EurekaServerApplicationTests.java
@@ -1,13 +1,0 @@
-package com.mst.eurekaserver;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class EurekaServerApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
