+++ backend/user-service/src/main/java/com/caffeinedoctor/userservice/UserServiceApplication.java
@@ -1,0 +1,15 @@
+package com.caffeinedoctor.userservice;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.cloud.client.discovery.EnableDiscoveryClient;
+
+@SpringBootApplication
+@EnableDiscoveryClient // 유레카 클라이언트로 등록
+public class UserServiceApplication {
+
+	public static void main(String[] args) {
+		SpringApplication.run(UserServiceApplication.class, args);
+	}
+
+}
