+++ backend/user-service/src/main/java/com/caffeinedoctor/userservice/controller/UserController.java
@@ -1,0 +1,13 @@
+package com.caffeinedoctor.userservice.controller;
+
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.RestController;
+
+@RestController
+public class UserController {
+
+    @GetMapping("/")
+    public String getServiceName(){
+        return "user-service";
+    }
+}
