--- backend/user-service/src/main/java/com/mst/userservice/controller/UserController.java
@@ -1,13 +1,0 @@
-package com.mst.userservice.controller;
-
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.RestController;
-
-@RestController
-public class UserController {
-
-    @GetMapping("/")
-    public String getServiceName(){
-        return "user-service";
-    }
-}
