+++ backend/user-service/src/main/resources/application.yml
@@ -1,19 +1,19 @@
-# ?? ????? ?? ?? ??
+# 현재 프로젝트 포트 지정
 server:
   port: 8081
 
-# ?? ????? ?? ??
+# 현재 프로젝트 이름 지정
 spring:
   application:
     name: user-service
 
-# ?? ????? eureka ??
+# eureka 설정
 eureka:
   client:
-    # eureka ??? ?? ??
+    # eureka 등록 여부
     register-with-eureka: true
-    # eureka ???? ??? ??? ??? ???? ??
+    # eureka 서버로 부터 인스턴스들의 정보를 가져올 것인지 설정하는 속성
     fetch-registry: true
-    # ??? ??? ??????? ??? ??
+    # eureka 서버 주소에 마이크로서비스 정보 등록
     service-url:
       defaultZone: http://admin:1234mst@localhost:8761/eureka
