+++ backend/user-service/src/test/java/com/caffeinedoctor/userservice/UserServiceApplicationTests.java
@@ -1,0 +1,13 @@
+package com.caffeinedoctor.userservice;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class UserServiceApplicationTests {
+
+	@Test
+	void contextLoads() {
+	}
+
+}
