--- backend/user-service/src/test/java/com/mst/userservice/UserServiceApplicationTests.java
@@ -1,13 +1,0 @@
-package com.mst.userservice;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class UserServiceApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
